1. Image upload issues:
If you upload an image with the same exact title as of a previously uploaded image, it will get uploaded. But then, if you try to navigate to one of the images with the same title, the image uploader will display an error.
Please fix this issue, so that the application should not show an error and take you to respective image’s page.

Please see more details of this issue on Github.

https://github.com/upgrad-edu/Course_4_Assignment/issues/1